requirements.txt
pip-tools==6.5.1
pandas==2.0.0
google-cloud-batch==0.7.0
google-cloud-storage==2.8.0
google-cloud-workflows-1.10.1
click==8.1.3
docker==6.0.1
